Clean up dom utilities
