+++ packages/cfpb-design-system/cfpb-design-system.js
@@ -1,1 +1,1 @@
-!function(t){var e={};function n(i){if(e[i])return e[i].exports;var s=e[i]={i:i,l:!1,exports:{}};return t[i].call(s.exports,s,s.exports,n),s.l=!0,s.exports}n.m=t,n.c=e,n.d=function(t,e,i){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:i})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var i=Object.create(null);if(n.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var s in t)n.d(i,s,function(e){return t[e]}.bind(null,s));return i},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=4)}([function(t,e){t.exports={on:function(t,e){var n=this.events=this.events||{};return n[t]=this.events[t]||[],n[t].push(e),this},off:function(t){return this.events&&this.events[t]&&delete this.events[t],this},trigger:function(t){var e=this.events||{};if(!1===e.hasOwnProperty(t))return this;for(var n=0,i=e[t].length;n<i;n++)this.events[t][n].apply(this,arguments);return this}}},function(t,e){t.exports={closest:function(t,e){if("closest"in t)return t.closest(e);for(var n,i=t.matches||t.webkitMatchesSelector||t.mozMatchesSelector||t.msMatchesSelector;t;)if(i.bind(t)(e)?n=t:t=t.parentElement,n)return t;return null}}},function(t,e){t.exports={DIRECTIONS:{UP:0,RIGHT:1,DOWN:-1,LEFT:-2},NO_OP_FUNCTION:function(){},PREFIXES:{PAGE:"p-",TEMPLATE:"t-",ORGANISM:"o-",MOLECULE:"m-",ATOM:"a-"},TYPES:{PAGE:1,TEMPLATE:2,ORGANISM:3,MOLECULE:4,ATOM:5},UNDEFINED:void 0}},function(t,e,n){var i=n(7),s=n(2).TYPES,r=i.extend({TYPE:s.ORGANISM,CHILD_TYPES:[s.MOLECULE,s.ATOM]});t.exports=r},function(t,e,n){t.exports=n(5)},function(t,e,n){var i=n(6),s=n(14);i.init(),s.init()},function(t,e,n){var i=n(1).closest,s=n(0),r=n(3),o=n(12),a=r.extend({ui:{base:".o-expandable",target:".o-expandable_target",content:".o-expandable_content",header:".o-expandable_header",label:".o-expandable_label"},classes:{targetExpanded:"o-expandable_target__expanded",targetCollapsed:"o-expandable_target__collapsed",group:"o-expandable-group",groupAccordion:"o-expandable-group__accordion"},events:{"click .o-expandable_target":"expandableClickHandler"},transition:null,isAccordionGroup:!1,activeAccordion:!1,initialize:function(){var t=new o(this.ui.content);this.transition=t.init(),this.transition.addEventListener("expandBegin",c.bind(this)),this.transition.addEventListener("collapseEnd",h.bind(this)),this.ui.content.classList.contains(o.CLASSES.EXPANDED)?this.ui.target.classList.add(this.classes.targetExpanded):(this.ui.target.classList.add(this.classes.targetCollapsed),this.ui.content.classList.add("u-hidden"));var e=i(this.ui.target,"."+this.classes.group);this.isAccordionGroup=null!==e&&e.classList.contains(this.classes.groupAccordion),this.isAccordionGroup&&s.on("accordionActivated",l.bind(this))},expandableClickHandler:function(){this.transition.toggleExpandable(),this.toggleTargetState(this.ui.target),this.isAccordionGroup&&(this.activeAccordion?this.activeAccordion=!1:(s.trigger("accordionActivated",{target:this}),this.activeAccordion=!0))},toggleTargetState:function(t){t.classList.contains(this.classes.targetExpanded)?(this.ui.target.classList.add(this.classes.targetCollapsed),this.ui.target.classList.remove(this.classes.targetExpanded)):(this.ui.target.classList.add(this.classes.targetExpanded),this.ui.target.classList.remove(this.classes.targetCollapsed))},getLabelText:function(){return this.ui.label.textContent.trim()}});function l(){this.activeAccordion&&(this.transition.toggleExpandable(),this.toggleTargetState(this.ui.target),this.activeAccordion=!1)}function c(){this.ui.content.classList.remove("u-hidden")}function h(){this.ui.content.classList.add("u-hidden")}t.exports=a},function(t,e,n){var i=n(8).assign,s=n(9).Delegate,r=n(0),o=n(11).isFunction;function a(t,e){this.element=t,this.initializers=[],this.uId=this.uniqueId("ac"),i(this,e),this.processModifiers(),this.ensureElement(),this.setCachedElements(),this.initializers.push(this.initialize),this.initializers.forEach((function(t){o(t)&&t.apply(this,arguments)}),this),this.trigger("component:initialized")}i(a.prototype,r,{tagName:"div",processModifiers:function(){this.modifiers&&this.modifiers.forEach((function(t){var e=t.ui.base.substring(1);this.element.classList.contains(e)&&(t.initialize&&(this.initializers.push(t.initialize),delete t.initialize),i(this,t))}),this)},render:function(){return this},ensureElement:function(){if(this.element)this.setElement(this.element);else{var t=i({},this.attributes);t.id=this.id||this.u_id,this.className&&(t.class=this.className),this.setElement(document.createElement(this.tagName)),this.setElementAttributes(t)}this.element.setAttribute("data-bound",!0)},setElement:function(t){return this.element&&this.undelegateEvents(),this.element=t,this.delegateEvents(),this},setCachedElements:function(){var t,e,n=i({},this.ui);for(t in n)n.hasOwnProperty(t)&&(1===(e=this.element.querySelectorAll(n[t])).length?n[t]=e[0]:e.length>1?n[t]=e:n[t]=null);return this.ui=n,n},destroy:function(){return this.element&&(this.element.parentNode.removeChild(this.element),this.element.view&&delete this.element.view,delete this.element),this.undelegateEvents(),this.trigger("component:destroyed"),!0},setElementAttributes:function(t){var e;for(e in t)t.hasOwnProperty(e)&&this.element.setAttribute(e,t[e])},delegateEvents:function(t){var e,n,i,r=/^(\S+)\s*(.*)$/;if(!(t=t||(t=this.events)))return this;for(e in this.undelegateEvents(),this._delegate=new s(this.element),t)({}).hasOwnProperty.call(t,e)&&(n=t[e],o(this[n])&&(n=this[n]),n&&(i=e.match(r),this.delegate(i[1],i[2],n.bind(this))));return this.trigger("component:bound"),this},delegate:function(t,e,n){return this._delegate.on(t,e,n),this},undelegateEvents:function(){return this._delegate&&this._delegate.destroy(),this.element.removeAttribute("data-bound"),this},uniqueId:function(t){return t+"_"+Math.random().toString(36).substr(2,9)}}),a.extend=function(t){function e(){return this._super=a.prototype,a.apply(this,arguments)}return e.prototype=Object.create(a.prototype),i(e.prototype,t),i(e,a),t.hasOwnProperty("ui")&&t.ui.hasOwnProperty("base")&&(e.selector=t.ui.base),e.constants={},e},a.init=function(t){for(var e,n=(t||document).querySelectorAll(this.selector),i=[],s=0,r=n.length;s<r;s++)!1===(e=n[s]).hasAttribute("data-bound")&&i.push(new this(e));return i},t.exports=a},function(t,e){function n(t){return"[object Object]"===Object.prototype.toString.call(t)}t.exports={assign:function t(e){e=e||{};for(var i=1,s=arguments.length;i<s;i++){var r=arguments[i]||{},o=void 0;for(o in r)if(Object.prototype.hasOwnProperty.call(r,o)){var a=r[o];n(a)?t(e[o]||(e[o]={}),a):e[o]=a}}return e}}},function(t,e,n){"use strict";var i=n(10);t.exports=function(t){return new i(t)},t.exports.Delegate=i},function(t,e,n){"use strict";function i(t){this.listenerMap=[{},{}],t&&this.root(t),this.handle=i.prototype.handle.bind(this),this._removedListeners=[]}t.exports=i,i.prototype.root=function(t){var e,n=this.listenerMap;if(this.rootElement){for(e in n[1])n[1].hasOwnProperty(e)&&this.rootElement.removeEventListener(e,this.handle,!0);for(e in n[0])n[0].hasOwnProperty(e)&&this.rootElement.removeEventListener(e,this.handle,!1)}if(!t||!t.addEventListener)return this.rootElement&&delete this.rootElement,this;for(e in this.rootElement=t,n[1])n[1].hasOwnProperty(e)&&this.rootElement.addEventListener(e,this.handle,!0);for(e in n[0])n[0].hasOwnProperty(e)&&this.rootElement.addEventListener(e,this.handle,!1);return this},i.prototype.captureForType=function(t){return-1!==["blur","error","focus","load","resize","scroll"].indexOf(t)},i.prototype.on=function(t,e,n,i){var l,c,h,u;if(!t)throw new TypeError("Invalid event type: "+t);if("function"==typeof e&&(i=n,n=e,e=null),void 0===i&&(i=this.captureForType(t)),"function"!=typeof n)throw new TypeError("Handler must be a type of Function");return l=this.rootElement,(c=this.listenerMap[i?1:0])[t]||(l&&l.addEventListener(t,this.handle,i),c[t]=[]),e?/^[a-z]+$/i.test(e)?(u=e,h=r):/^#[a-z0-9\-_]+$/i.test(e)?(u=e.slice(1),h=a):(u=e,h=s):(u=null,h=o.bind(this)),c[t].push({selector:e,handler:n,matcher:h,matcherParam:u}),this},i.prototype.off=function(t,e,n,i){var s,r,o,a,l;if("function"==typeof e&&(i=n,n=e,e=null),void 0===i)return this.off(t,e,n,!0),this.off(t,e,n,!1),this;if(o=this.listenerMap[i?1:0],!t){for(l in o)o.hasOwnProperty(l)&&this.off(l,e,n);return this}if(!(a=o[t])||!a.length)return this;for(s=a.length-1;s>=0;s--)r=a[s],e&&e!==r.selector||n&&n!==r.handler||(this._removedListeners.push(r),a.splice(s,1));return a.length||(delete o[t],this.rootElement&&this.rootElement.removeEventListener(t,this.handle,i)),this},i.prototype.handle=function(t){var e,n,i,s,r,o=t.type,a=[];if(!0!==t.ftLabsDelegateIgnore){switch(3===(r=t.target).nodeType&&(r=r.parentNode),r.correspondingUseElement&&(r=r.correspondingUseElement),i=this.rootElement,t.eventPhase||(t.target!==t.currentTarget?3:2)){case 1:a=this.listenerMap[1][o];break;case 2:this.listenerMap[0]&&this.listenerMap[0][o]&&(a=a.concat(this.listenerMap[0][o])),this.listenerMap[1]&&this.listenerMap[1][o]&&(a=a.concat(this.listenerMap[1][o]));break;case 3:a=this.listenerMap[0][o]}var l,c=[];for(n=a.length;r&&n;){for(e=0;e<n&&(s=a[e]);e++)r.tagName&&["button","input","select","textarea"].indexOf(r.tagName.toLowerCase())>-1&&r.hasAttribute("disabled")?c=[]:s.matcher.call(r,s.matcherParam,r)&&c.push([t,r,s]);if(r===i)break;if(n=a.length,(r=r.parentElement||r.parentNode)instanceof HTMLDocument)break}for(e=0;e<c.length;e++)if(!(this._removedListeners.indexOf(c[e][2])>-1)&&!1===this.fire.apply(this,c[e])){c[e][0].ftLabsDelegateIgnore=!0,c[e][0].preventDefault(),l=!1;break}return l}},i.prototype.fire=function(t,e,n){return n.handler.call(e,t,e)};var s=function(t){if(t){var e=t.prototype;return e.matches||e.matchesSelector||e.webkitMatchesSelector||e.mozMatchesSelector||e.msMatchesSelector||e.oMatchesSelector}}(Element);function r(t,e){return t.toLowerCase()===e.tagName.toLowerCase()}function o(t,e){return this.rootElement===window?e===document||e===document.documentElement||e===window:this.rootElement===e}function a(t,e){return t===e.id}i.prototype.destroy=function(){this.off(),this.root()}},function(t,e){function n(t){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}var i=Object.prototype.toString;function s(t){return void 0===t}function r(t){return"[object String]"===i.call(t)}var o=Array.isArray||function(t){return"[object Array]"===i.call(t)};t.exports={isUndefined:s,isDefined:function(t){return void 0!==t},isObject:function(t){return null!==t&&"object"===n(t)},isString:r,isNumber:function(t){return"[object Number]"===i.call(t)},isDate:function(t){return"[object Date]"===i.call(t)},isArray:o,isFunction:function(t){return"[object Function]"===i.call(t)},isEmpty:function(t){return s(t)||null===t||r(t)&&t.length<=0||/^\s*$/.test(t)}}},function(t,e,n){var i=n(0),s=n(13),r={BASE_CLASS:"o-expandable_content__transition",EXPANDED:"o-expandable_content__expanded",COLLAPSED:"o-expandable_content__collapsed",OPEN_DEFAULT:"o-expandable_content__onload-open"};function o(t){var e,n=new s(t,r);function o(){t.classList.contains(r.EXPANDED)?(this.dispatchEvent("expandEnd",{target:this}),t.scrollHeight>e&&(t.style.maxHeight=t.scrollHeight+"px")):t.classList.contains(r.COLLAPSED)&&this.dispatchEvent("collapseEnd",{target:this})}return this.addEventListener=i.on,this.dispatchEvent=i.trigger,this.removeEventListener=i.off,this.animateOff=n.animateOff,this.animateOn=n.animateOn,this.halt=n.halt,this.isAnimated=n.isAnimated,this.setElement=n.setElement,this.remove=n.remove,this.init=function(){return n.init(),n.addEventListener(s.END_EVENT,o.bind(this)),t.classList.contains(r.OPEN_DEFAULT)?this.expand():this.collapse(),this},this.toggleExpandable=function(){return t.classList.contains(r.COLLAPSED)?this.expand():this.collapse(),this},this.collapse=function(){return this.dispatchEvent("collapseBegin",{target:this}),e=t.scrollHeight,t.style.maxHeight="0",n.applyClass(r.COLLAPSED),this},this.expand=function(){return this.dispatchEvent("expandBegin",{target:this}),(!e||t.scrollHeight>e)&&(e=t.scrollHeight),t.style.maxHeight=e+"px",n.applyClass(r.EXPANDED),this},this}o.CLASSES=r,t.exports=o},function(t,e,n){var i=n(0);function s(t,e){var n,r,o,a,l,c=e,h=!1,u=!1;function d(t){n&&(v(),p()),(n=t).classList.add(c.BASE_CLASS),o=function(t){if(!t){throw new Error("Element does not have TransitionEnd event. It may be null!")}var e,n,i={WebkitTransition:"webkitTransitionEnd",MozTransition:"transitionend",OTransition:"oTransitionEnd otransitionend",transition:"transitionend"};for(n in i)if(i.hasOwnProperty(n)&&void 0!==t.style[n]){e=i[n];break}return e}(n)}function p(){return n?(n.classList.remove(s.NO_ANIMATION_CLASS),this):this}function f(){h&&(n.style.webkitTransitionDuration="0",n.style.mozTransitionDuration="0",n.style.oTransitionDuration="0",n.style.transitionDuration="0",n.removeEventListener(o,a),a(),n.style.webkitTransitionDuration="",n.style.mozTransitionDuration="",n.style.oTransitionDuration="",n.style.transitionDuration="")}function m(){h=!0,o?(n.addEventListener(o,a),this.trigger(s.BEGIN_EVENT,{target:this})):(this.trigger(s.BEGIN_EVENT,{target:this}),a())}function g(){n.removeEventListener(o,a)}function E(){g(),this.trigger(s.END_EVENT,{target:this}),h=!1}function b(){var t;for(t in c)c.hasOwnProperty(t)&&c[t]!==c.BASE_CLASS&&n.classList.contains(c[t])&&n.classList.remove(c[t])}function v(){return!!n&&(f(),n.classList.remove(c.BASE_CLASS),b(),!0)}return this.addEventListener=i.on,this.trigger=i.trigger,this.removeEventListener=i.off,this.animateOff=function(){return n?(n.classList.add(s.NO_ANIMATION_CLASS),this):this},this.animateOn=p,this.applyClass=function(t){return!!n&&(u||(b(),u=!0),!n.classList.contains(t)&&(g(),n.classList.remove(r),r=t,l(),n.classList.add(r),!0))},this.halt=f,this.init=function(){return a=E.bind(this),l=m.bind(this),d(t),this},this.isAnimated=function(){return!!n&&!n.classList.contains(s.NO_ANIMATION_CLASS)},this.remove=v,this.setElement=d,this}s.BEGIN_EVENT="transitionBegin",s.END_EVENT="transitionEnd",s.NO_ANIMATION_CLASS="u-no-animation",t.exports=s},function(t,e,n){var i=n(2),s=n(3),r=n(15),o=n(16),a=s.extend({ui:{base:".o-table"},modifiers:[r,o]});a.constants.DIRECTIONS=i.DIRECTIONS,t.exports=a},function(t,e,n){var i=n(2),s=n(1).closest,r=i.DIRECTIONS,o=i.UNDEFINED,a={ui:{base:".o-table__sortable",tableBody:"tbody",sortButton:".sorted-up, .sorted-down"},classes:{sortDown:"sorted-down",sortUp:"sorted-up"},events:{"click .sortable":"onSortableClick"},initialize:function(){this.sortClass=o,this.sortColumnIndex=o,this.sortDirection=o,this.tableData=[],this.bindProperties(),this.ui.sortButton&&(this.sortColumnIndex=this.getColumnIndex(),this.sortDirection=r.UP,this.ui.sortButton.classList.contains(this.classes.sortDown)&&(this.sortDirection=r.DOWN),this.updateTable())},bindProperties:function(){var t;Object.defineProperty(this,"sortDirection",{configurable:!0,get:function(){return t},set:function(e){e===r.UP?this.sortClass=this.classes.sortUp:e===r.DOWN&&(this.sortClass=this.classes.sortDown),t=e}})},getColumnIndex:function(t){return s(t||this.ui.sortButton,"td, th").cellIndex},updateTable:function(){return this.updateTableData()&&this.updateTableDom()},updateTableData:function(t){var e,n=this.ui.tableBody.querySelectorAll("tr");this.tableData=[],t=t||this.sortColumnIndex;for(var i=0,s=n.length;i<s;++i)(e=n[i].cells[t])&&(e=e.textContent.trim()),this.tableData.push([e,n[i]]);var r=this.ui.sortButton.getAttribute("data-sort_type");return this.tableData.sort(this.tableDataSorter(this.sortDirection,r)),this.tableData},updateTableDom:function(){var t=this.ui.tableBody;for(;t.lastChild;)t.removeChild(t.lastChild);for(var e=document.createDocumentFragment(),n=0;n<this.tableData.length;n++)e.appendChild(this.tableData[n][1]);return t.appendChild(e),this.trigger("table:updated"),t},tableDataSorter:function(t,e){return function(n,i){var s=t===r.DOWN?-1:1,o=0,a=/[^\d.-]/g;return n=n[0],i=i[0],"number"===e&&(n=Number(n.replace(a,"")),i=Number(i.replace(a,""))),n<i?o=-1*s:n>i&&(o=s),o}},onSortableClick:function(t){this.ui.sortButton&&this.ui.sortButton.classList.remove(this.sortClass);this.ui.sortButton===t.target?this.sortDirection=~this.sortDirection:(this.ui.sortButton=t.target,this.sortColumnIndex=this.getColumnIndex(),this.sortDirection=r.UP);return this.ui.sortButton.classList.add(this.sortClass),this.updateTable(),this}};t.exports=a},function(t,e,n){var i=n(1).closest,s={ui:{base:".o-table__row-links"},events:{"click tbody tr":"onRowLinkClick"},onRowLinkClick:function(t){var e=t.target;if(e&&"A"===e.tagName)return;var n=(e=i(t.target,"tr")).querySelector("a");n&&(window.location=n.getAttribute("href"))}};t.exports=s}]);
+!function(){var t,e,n={852:function(t){"use strict";function e(t){this.listenerMap=[{},{}],t&&this.root(t),this.handle=e.prototype.handle.bind(this),this._removedListeners=[]}t.exports=e,e.prototype.root=function(t){var e,n=this.listenerMap;if(this.rootElement){for(e in n[1])n[1].hasOwnProperty(e)&&this.rootElement.removeEventListener(e,this.handle,!0);for(e in n[0])n[0].hasOwnProperty(e)&&this.rootElement.removeEventListener(e,this.handle,!1)}if(!t||!t.addEventListener)return this.rootElement&&delete this.rootElement,this;for(e in this.rootElement=t,n[1])n[1].hasOwnProperty(e)&&this.rootElement.addEventListener(e,this.handle,!0);for(e in n[0])n[0].hasOwnProperty(e)&&this.rootElement.addEventListener(e,this.handle,!1);return this},e.prototype.captureForType=function(t){return-1!==["blur","error","focus","load","resize","scroll"].indexOf(t)},e.prototype.on=function(t,e,o,a){var l,h,c,u;if(!t)throw new TypeError("Invalid event type: "+t);if("function"==typeof e&&(a=o,o=e,e=null),void 0===a&&(a=this.captureForType(t)),"function"!=typeof o)throw new TypeError("Handler must be a type of Function");return l=this.rootElement,(h=this.listenerMap[a?1:0])[t]||(l&&l.addEventListener(t,this.handle,a),h[t]=[]),e?/^[a-z]+$/i.test(e)?(u=e,c=i):/^#[a-z0-9\-_]+$/i.test(e)?(u=e.slice(1),c=r):(u=e,c=n):(u=null,c=s.bind(this)),h[t].push({selector:e,handler:o,matcher:c,matcherParam:u}),this},e.prototype.off=function(t,e,n,i){var s,r,o,a,l;if("function"==typeof e&&(i=n,n=e,e=null),void 0===i)return this.off(t,e,n,!0),this.off(t,e,n,!1),this;if(o=this.listenerMap[i?1:0],!t){for(l in o)o.hasOwnProperty(l)&&this.off(l,e,n);return this}if(!(a=o[t])||!a.length)return this;for(s=a.length-1;s>=0;s--)r=a[s],e&&e!==r.selector||n&&n!==r.handler||(this._removedListeners.push(r),a.splice(s,1));return a.length||(delete o[t],this.rootElement&&this.rootElement.removeEventListener(t,this.handle,i)),this},e.prototype.handle=function(t){var e,n,i,s,r,o=t.type,a=[],l="ftLabsDelegateIgnore";if(!0!==t[l]){switch(3===(r=t.target).nodeType&&(r=r.parentNode),r.correspondingUseElement&&(r=r.correspondingUseElement),i=this.rootElement,t.eventPhase||(t.target!==t.currentTarget?3:2)){case 1:a=this.listenerMap[1][o];break;case 2:this.listenerMap[0]&&this.listenerMap[0][o]&&(a=a.concat(this.listenerMap[0][o])),this.listenerMap[1]&&this.listenerMap[1][o]&&(a=a.concat(this.listenerMap[1][o]));break;case 3:a=this.listenerMap[0][o]}var h,c=[];for(n=a.length;r&&n;){for(e=0;e<n&&(s=a[e]);e++)r.tagName&&["button","input","select","textarea"].indexOf(r.tagName.toLowerCase())>-1&&r.hasAttribute("disabled")?c=[]:s.matcher.call(r,s.matcherParam,r)&&c.push([t,r,s]);if(r===i)break;if(n=a.length,(r=r.parentElement||r.parentNode)instanceof HTMLDocument)break}for(e=0;e<c.length;e++)if(!(this._removedListeners.indexOf(c[e][2])>-1)&&!1===this.fire.apply(this,c[e])){c[e][0][l]=!0,c[e][0].preventDefault(),h=!1;break}return h}},e.prototype.fire=function(t,e,n){return n.handler.call(e,t,e)};var n=function(t){if(t){var e=t.prototype;return e.matches||e.matchesSelector||e.webkitMatchesSelector||e.mozMatchesSelector||e.msMatchesSelector||e.oMatchesSelector}}(Element);function i(t,e){return t.toLowerCase()===e.tagName.toLowerCase()}function s(t,e){return this.rootElement===window?e===document||e===document.documentElement||e===window:this.rootElement===e}function r(t,e){return t===e.id}e.prototype.destroy=function(){this.off(),this.root()}},161:function(t,e,n){"use strict";var i=n(852);t.exports=function(t){return new i(t)},t.exports.Delegate=i},335:function(t,e,n){"use strict";function i(t){return"[object Object]"===Object.prototype.toString.call(t)}function s(t){t=t||{};for(var e=1,n=arguments.length;e<n;e++){var r=arguments[e]||{},o=void 0;for(o in r)if(Object.prototype.hasOwnProperty.call(r,o)){var a=r[o];i(a)?s(t[o]||(t[o]={}),a):t[o]=a}}return t}n.d(e,{Z:function(){return h}});var r=n(147),o=n(814),a=n(161).Delegate;function l(t,e){this.element=t,this.initializers=[],this.uId=this.uniqueId("ac"),s(this,e),this.processModifiers(),this.ensureElement(),this.setCachedElements(),this.initializers.push(this.initialize),this.initializers.forEach((function(t){o.Z.isFunction(t)&&t.apply(this,arguments)}),this),this.dispatchEvent("component:initialized")}s(l.prototype,new r.Z,{tagName:"div",processModifiers:function(){this.modifiers&&this.modifiers.forEach((function(t){var e=t.ui.base.substring(1);this.element.classList.contains(e)&&(t.initialize&&(this.initializers.push(t.initialize),delete t.initialize),s(this,t))}),this)},render:function(){return this},ensureElement:function(){if(this.element)this.setElement(this.element);else{var t=s({},this.attributes);t.id=this.id||this.u_id,this.className&&(t.class=this.className),this.setElement(document.createElement(this.tagName)),this.setElementAttributes(t)}this.element.setAttribute("data-bound",!0)},setElement:function(t){return this.element&&this.undelegateEvents(),this.element=t,this.delegateEvents(),this},setCachedElements:function(){var t,e,n=s({},this.ui);for(t in n)n.hasOwnProperty(t)&&(1===(e=this.element.querySelectorAll(n[t])).length?n[t]=e[0]:e.length>1?n[t]=e:n[t]=null);return this.ui=n,n},destroy:function(){return this.element&&(this.element.parentNode.removeChild(this.element),this.element.view&&delete this.element.view,delete this.element),this.undelegateEvents(),this.dispatchEvent("component:destroyed"),!0},setElementAttributes:function(t){var e;for(e in t)t.hasOwnProperty(e)&&this.element.setAttribute(e,t[e])},delegateEvents:function(t){var e,n,i,s=/^(\S+)\s*(.*)$/;if(!(t=t||(t=this.events)))return this;for(e in this.undelegateEvents(),this._delegate=new a(this.element),t)({}).hasOwnProperty.call(t,e)&&(n=t[e],o.Z.isFunction(this[n])&&(n=this[n]),n&&(i=e.match(s),this.delegate(i[1],i[2],n.bind(this))));return this.dispatchEvent("component:bound"),this},delegate:function(t,e,n){return this._delegate.on(t,e,n),this},undelegateEvents:function(){return this._delegate&&this._delegate.destroy(),this.element.removeAttribute("data-bound"),this},uniqueId:function(t){return t+"_"+Math.random().toString(36).substr(2,9)}}),l.extend=function(t){function e(){return this._super=l.prototype,l.apply(this,arguments)}return e.prototype=Object.create(l.prototype),s(e.prototype,t),s(e,l),t.hasOwnProperty("ui")&&t.ui.hasOwnProperty("base")&&(e.selector=t.ui.base),e.constants={},e},l.init=function(t){for(var e,n=(t||document).querySelectorAll(this.selector),i=[],s=0,r=n.length;s<r;s++)!1===(e=n[s]).hasAttribute("data-bound")&&i.push(new this(e));return i};var h=l},147:function(t,e){"use strict";e.Z=function(){var t={};return this.addEventListener=function(e,n){return t.hasOwnProperty(e)?t[e].push(n):t[e]=[n],this},this.removeEventListener=function(e,n){if(!t.hasOwnProperty(e))return this;var i=t[e].indexOf(n);return-1!==i&&t[e].splice(i,1),this},this.dispatchEvent=function(e,n){if(!t.hasOwnProperty(e))return this;n=n||{};for(var i=t[e],s=0,r=i.length;s<r;s++)i[s].call(this,n);return this},this.getRegisteredEvents=function(){return t},this}},154:function(t,e,n){"use strict";n.d(e,{o:function(){return i}});n(814);function i(t,e){if("closest"in t)return t.closest(e);for(var n,i=t.matches||t.webkitMatchesSelector||t.mozMatchesSelector||t.msMatchesSelector;t;)if(i.bind(t)(e)?n=t:t=t.parentElement,n)return t;return null}},814:function(t,e){"use strict";function n(t){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}var i=Object.prototype.toString;function s(t){return void 0===t}function r(t){return"[object String]"===i.call(t)}var o=Array.isArray||function(t){return"[object Array]"===i.call(t)};e.Z={isUndefined:s,isDefined:function(t){return void 0!==t},isObject:function(t){return null!==t&&"object"===n(t)},isString:r,isNumber:function(t){return"[object Number]"===i.call(t)},isDate:function(t){return"[object Date]"===i.call(t)},isArray:o,isFunction:function(t){return"[object Function]"===i.call(t)},isEmpty:function(t){return s(t)||null===t||r(t)&&t.length<=0||/^\s*$/.test(t)}}},270:function(t,e,n){"use strict";n.r(e),n.d(e,{default:function(){return m}});var i=n(154),s=n(147),r=n(335);function o(t,e){var n,i,r,a,l,h=e,c=!1,u=!1;function d(t){n&&(g(),p()),(n=t).classList.add(h.BASE_CLASS),r=function(t){if(!t){throw new Error("Element does not have TransitionEnd event. It may be null!")}var e,n,i={WebkitTransition:"webkitTransitionEnd",MozTransition:"transitionend",OTransition:"oTransitionEnd otransitionend",transition:"transitionend"};for(n in i)if(i.hasOwnProperty(n)&&void 0!==t.style[n]){e=i[n];break}return e}(n)}function p(){return n?(n.classList.remove(o.NO_ANIMATION_CLASS),this):this}function f(){c&&(n.style.webkitTransitionDuration="0",n.style.mozTransitionDuration="0",n.style.oTransitionDuration="0",n.style.transitionDuration="0",n.removeEventListener(r,a),a(),n.style.webkitTransitionDuration="",n.style.mozTransitionDuration="",n.style.oTransitionDuration="",n.style.transitionDuration="")}function m(){c=!0,r?(n.addEventListener(r,a),this.dispatchEvent(o.BEGIN_EVENT,{target:this})):(this.dispatchEvent(o.BEGIN_EVENT,{target:this}),a())}function v(){n.removeEventListener(r,a)}function E(){v(),this.dispatchEvent(o.END_EVENT,{target:this}),c=!1}function b(){var t;for(t in h)h.hasOwnProperty(t)&&h[t]!==h.BASE_CLASS&&n.classList.contains(h[t])&&n.classList.remove(h[t])}function g(){return!!n&&(f(),n.classList.remove(h.BASE_CLASS),b(),!0)}var y=new s.Z;return this.addEventListener=y.addEventListener,this.dispatchEvent=y.dispatchEvent,this.removeEventListener=y.removeEventListener,this.animateOff=function(){return n?(n.classList.add(o.NO_ANIMATION_CLASS),this):this},this.animateOn=p,this.applyClass=function(t){return!!n&&(u||(b(),u=!0),!n.classList.contains(t)&&(v(),n.classList.remove(i),i=t,l(),n.classList.add(i),!0))},this.halt=f,this.init=function(){return a=E.bind(this),l=m.bind(this),d(t),this},this.isAnimated=function(){return!!n&&!n.classList.contains(o.NO_ANIMATION_CLASS)},this.remove=g,this.setElement=d,this}o.BEGIN_EVENT="transitionBegin",o.END_EVENT="transitionEnd",o.NO_ANIMATION_CLASS="u-no-animation";var a=o,l={BASE_CLASS:"o-expandable_content__transition",EXPANDED:"o-expandable_content__expanded",COLLAPSED:"o-expandable_content__collapsed",OPEN_DEFAULT:"o-expandable_content__onload-open"};function h(t){var e,n=new a(t,l);function i(){t.classList.contains(l.EXPANDED)?(this.dispatchEvent("expandEnd",{target:this}),t.scrollHeight>e&&(t.style.maxHeight=t.scrollHeight+"px")):t.classList.contains(l.COLLAPSED)&&this.dispatchEvent("collapseEnd",{target:this})}var r=new s.Z;return this.addEventListener=r.addEventListener,this.dispatchEvent=r.dispatchEvent,this.removeEventListener=r.removeEventListener,this.animateOff=n.animateOff,this.animateOn=n.animateOn,this.halt=n.halt,this.isAnimated=n.isAnimated,this.setElement=n.setElement,this.remove=n.remove,this.init=function(){return n.init(),n.addEventListener(a.END_EVENT,i.bind(this)),t.classList.contains(l.OPEN_DEFAULT)?this.expand():this.collapse(),this},this.toggleExpandable=function(){return t.classList.contains(l.COLLAPSED)?this.expand():this.collapse(),this},this.collapse=function(){return this.dispatchEvent("collapseBegin",{target:this}),e=t.scrollHeight,t.style.maxHeight="0",n.applyClass(l.COLLAPSED),this},this.expand=function(){return this.dispatchEvent("expandBegin",{target:this}),(!e||t.scrollHeight>e)&&(e=t.scrollHeight),t.style.maxHeight=e+"px",n.applyClass(l.EXPANDED),this},this}h.CLASSES=l;var c=h,u=new s.Z;function d(){this.activeAccordion&&(this.transition.toggleExpandable(),this.toggleTargetState(this.ui.target),this.activeAccordion=!1)}function p(){this.ui.content.classList.remove("u-hidden")}function f(){this.ui.content.classList.add("u-hidden")}var m=r.Z.extend({ui:{base:".o-expandable",target:".o-expandable_target",content:".o-expandable_content",header:".o-expandable_header",label:".o-expandable_label"},classes:{targetExpanded:"o-expandable_target__expanded",targetCollapsed:"o-expandable_target__collapsed",group:"o-expandable-group",groupAccordion:"o-expandable-group__accordion"},events:{"click .o-expandable_target":"expandableClickHandler"},transition:null,isAccordionGroup:!1,activeAccordion:!1,initialize:function(){var t=new c(this.ui.content);this.transition=t.init(),this.transition.addEventListener("expandBegin",p.bind(this)),this.transition.addEventListener("collapseEnd",f.bind(this)),this.ui.content.classList.contains(c.CLASSES.EXPANDED)?this.ui.target.classList.add(this.classes.targetExpanded):(this.ui.target.classList.add(this.classes.targetCollapsed),this.ui.content.classList.add("u-hidden"));var e=(0,i.o)(this.ui.target,"."+this.classes.group);this.isAccordionGroup=null!==e&&e.classList.contains(this.classes.groupAccordion),this.isAccordionGroup&&u.addEventListener("accordionActivated",d.bind(this))},expandableClickHandler:function(){this.transition.toggleExpandable(),this.toggleTargetState(this.ui.target),this.isAccordionGroup&&(this.activeAccordion?this.activeAccordion=!1:(u.dispatchEvent("accordionActivated",{target:this}),this.activeAccordion=!0))},toggleTargetState:function(t){t.classList.contains(this.classes.targetExpanded)?(this.ui.target.classList.add(this.classes.targetCollapsed),this.ui.target.classList.remove(this.classes.targetExpanded)):(this.ui.target.classList.add(this.classes.targetExpanded),this.ui.target.classList.remove(this.classes.targetCollapsed))},getLabelText:function(){return this.ui.label.textContent.trim()}})},518:function(t,e,n){"use strict";n.r(e),n.d(e,{default:function(){return d}});var i={UP:0,RIGHT:1,DOWN:-1,LEFT:-2},s=undefined,r=n(335),o=n(154),a=i,l=s;var h={ui:{base:".o-table__sortable",tableBody:"tbody",sortButton:".sorted-up, .sorted-down"},classes:{sortDown:"sorted-down",sortUp:"sorted-up"},events:{"click .sortable":"onSortableClick"},initialize:function(){this.sortClass=l,this.sortColumnIndex=l,this.sortDirection=l,this.tableData=[],this.bindProperties(),this.ui.sortButton&&(this.sortColumnIndex=this.getColumnIndex(),this.sortDirection=a.UP,this.ui.sortButton.classList.contains(this.classes.sortDown)&&(this.sortDirection=a.DOWN),this.updateTable())},bindProperties:function(){var t;Object.defineProperty(this,"sortDirection",{configurable:!0,get:function(){return t},set:function(e){e===a.UP?this.sortClass=this.classes.sortUp:e===a.DOWN&&(this.sortClass=this.classes.sortDown),t=e}})},getColumnIndex:function(t){return(0,o.o)(t||this.ui.sortButton,"td, th").cellIndex},updateTable:function(){return this.updateTableData()&&this.updateTableDom()},updateTableData:function(t){var e,n=this.ui.tableBody.querySelectorAll("tr");this.tableData=[],t=t||this.sortColumnIndex;for(var i=0,s=n.length;i<s;++i)(e=n[i].cells[t])&&(e=e.textContent.trim()),this.tableData.push([e,n[i]]);var r=this.ui.sortButton.getAttribute("data-sort_type");return this.tableData.sort(this.tableDataSorter(this.sortDirection,r)),this.tableData},updateTableDom:function(){var t=this.ui.tableBody;for(;t.lastChild;)t.removeChild(t.lastChild);for(var e=document.createDocumentFragment(),n=0;n<this.tableData.length;n++)e.appendChild(this.tableData[n][1]);return t.appendChild(e),this.dispatchEvent("table:updated"),t},tableDataSorter:function(t,e){return function(n,i){var s=t===a.DOWN?-1:1,r=0,o=/[^\d.-]/g;return n=n[0],i=i[0],"number"===e&&(n=Number(n.replace(o,"")),i=Number(i.replace(o,""))),n<i?r=-1*s:n>i&&(r=s),r}},onSortableClick:function(t){this.ui.sortButton&&this.ui.sortButton.classList.remove(this.sortClass);this.ui.sortButton===t.target?this.sortDirection=~this.sortDirection:(this.ui.sortButton=t.target,this.sortColumnIndex=this.getColumnIndex(),this.sortDirection=a.UP);return this.ui.sortButton.classList.add(this.sortClass),this.updateTable(),this}};var c={ui:{base:".o-table__row-links"},events:{"click tbody tr":"onRowLinkClick"},onRowLinkClick:function(t){var e=t.target;if(e&&"A"===e.tagName)return;var n=(e=(0,o.o)(t.target,"tr")).querySelector("a");n&&(window.location=n.getAttribute("href"))}},u=r.Z.extend({ui:{base:".o-table"},modifiers:[h,c]});u.constants.DIRECTIONS=i;var d=u}},i={};function s(t){if(i[t])return i[t].exports;var e=i[t]={exports:{}};return n[t](e,e.exports,s),e.exports}s.d=function(t,e){for(var n in e)s.o(e,n)&&!s.o(t,n)&&Object.defineProperty(t,n,{enumerable:!0,get:e[n]})},s.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},s.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},t=s(270),e=s(518),t.init(),e.init()}();
